Kassa

+++ Shop-S-V/Toode.cs
@@ -9,7 +9,7 @@ namespace Shop_S_V
     public class Toode
     {
         public string Nimi { get; set; }
-        public int Hind { get; set; }
+        public double Hind { get; set; }
         public int Kogus { get; set; }
     }
 }
